requirements.txt
python-dotenv
fastapi
httpx
uvicorn
pydantic
sqlalchemy
redis
redis[hiredis]
pytest
black
pytest-asyncio
PyJWT
passlib


